auto-service-stop.sh: Add check to service handling

+++ auto-service-stop.sh
@@ -1,21 +1,32 @@
 #!/bin/bash
 
 # stop apache2
-sudo service apache2 stop
-echo "Stop apache2 server"
+CMD=`systemctl is-enabled apache2 2>&1 | grep Failed`
+if [ -z "$CMD" ]; then
+	sudo service apache2 stop
+	echo "Stop apache2 server"
+fi
 
 # stop dlna server
-sudo service minidlna stop
-echo "Stop dlna server"
+CMD=`systemctl is-enabled minidlna 2>&1 | grep Failed`
+if [ -z "$CMD" ]; then
+	sudo service minidlna stop
+	echo "Stop dlna server"
+fi
 
 # stop samba server
-sudo service smbd stop
-echo "Stop samba server"
+CMD=`systemctl is-enabled smbd 2>&1 | grep Failed`
+if [ -z "$CMD" ]; then
+	sudo service smbd stop
+	echo "Stop samba server"
+fi
 
 # stop torrent server
 TORRENT=$(ps -ef | grep qbit | grep nox | awk '{print $2}')
-sudo kill -9 $TORRENT
-echo "Stop torrent server"
+if [ -n "$TORRENT" ]; then
+	sudo kill -9 $TORRENT
+	echo "Stop torrent server"
+fi
 
 # umount hdd
 sudo fuser -ku /home/Share
